requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openai==0.2.6
pycodestyle==2.6.0
PyJWT==1.7.1
pyngrok==5.0.0
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
toml==0.10.2
twilio==6.50.1
urllib3==1.26.2
Werkzeug==1.0.1
